requirements.txt
openai~=1.45.0
weasyprint~=62.3
pypdf2~=3.0.1
reportlab~=4.2.2
PyMuPDFb==1.24.10
fuzzywuzzy==0.18.0
pdfplumber==0.11.4
pdfkit==1.0.0
PyMuPDF==1.24.10
python_Levenshtein==0.26.0
matplotlib==3.9.2
pybboxes==0.1.6
scikit-learn==1.5.2
ultralytics==8.2.100
comet-ml==3.46.0
datasets==3.0.1
PyLaTeX==1.4.2
opencv-python~=4.10.0.84
numpy~=1.26.4
tqdm~=4.66.5
pillow~=10.4.0